gtkflow: allow recursion if the user demands it
there is now a property of GtkFlow.Nodeview called allow_recursion
if set to true, the nodeview will omit all prevention measures
for cycles in the graph. The user is responsible for implementing
nodes that can handle backpropagating signals without causing endless
loops.

added a little python3 example on how to use it

 greetings to @maweki