gtkflow: implemented force-based layouting system
one can now ask the library to arrange nodes on the nodeview
based on the laws of the physics of magnetical fields.
unconnected nodes repel each other and connected nodes attract
each other. nodes always try to keep a distance of 30 pixels
to each other. these constraints usually result in pretty layouts.
be careful: the algorithm is not smart enough yet to take the left-
to right order of nodes into account. and does not know that it e.g
makes sense to place a source-only node on the left screenside [in
LTR-environments].

relates to #53