gflow: added before_linking signal
the before_linking signal is being triggered when something,
either a user via a nodeview or the application programatically,
tries to establich a connection between two docs via the Dock.link-
method.

The programmer can implement their own checks wheter this connection
is valid or not. if the programmer returns true, the connection will
be made. otherwise the link method will stop.

i sense that @benwaffle will need this sooner or later.