gflow: fixed unlinking behaviour of sources
when the user connected a source to multiple sinks and then
deleted the node containing the source, only one sink would
be correctly disconnected, leaving the others in some invali
state

the problem was that we removed elements from the list while
iterating over it.

this has been fixed by iterating over a copy of the sink list.