gtkflow: update nodeview alloc for rubberband
when the rubberband was dragged there was no automatic
update of the nodeviews requested size. in some
node constellations this lead to the rubberband only
being drawn to the lowest point of the lowest node
and the rightmost point of the rightmost node.
calling an additional allocate_minimum() changes this.