dockline_renderer: only increment if type matches
the get_dock_position-method increments the count of docks it iterated
over and then multiplies the count with the height of a single dock to
determine the position of a dock inside a node. for the dockline-renderer
we must increment only if the dock has the same type as the dock we are
looking for.