gtkflow: made set_node_renderer usable from vala/c
i had not thought about the constructor of the node renderers
being internal. this is blatantly ignored by the gi-bindings,
or at least by python as far as i tested.
as i've never tried using the feature with vala this went unnoticed
quite some time. now we can trigger those internal noderenderers
by utilizing the values of a enum that represent the internal node
renderers als well as a CUSTOM-entry that is being chosen when
one wants to pass a noderenderer that has been created in the
application code.